# openflare-server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OpenFlare Server 管理端与 Agent API 文档。",
        "title": "OpenFlare Server API",
        "contact": {},
        "version": "3.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/dns-accounts/": {
            "get": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DnsAccounts"
                ],
                "summary": "List DNS accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DnsAccounts"
                ],
                "summary": "Create DNS account",
                "parameters": [
                    {
                        "description": "DNS account payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.DnsAccountInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dns-accounts/{id}/delete": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DnsAccounts"
                ],
                "summary": "Delete DNS account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "DNS Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dns-accounts/{id}/update": {
            "post": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DnsAccounts"
                ],
                "summary": "Update DNS account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "DNS Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "DNS account payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.DnsAccountInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/flared/apply-log": {
            "post": {
                "security": [
                    {
                        "TunnelTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flared"
                ],
                "summary": "Report OpenFlared apply result",
                "parameters": [
                    {
                        "description": "Apply log payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ApplyLogPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/flared/config/active": {
            "get": {
                "security": [
                    {
                        "TunnelTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flared"
                ],
                "summary": "Get active tunnel config for OpenFlared",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/flared/heartbeat": {
            "post": {
                "security": [
                    {
                        "TunnelTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flared"
                ],
                "summary": "Report OpenFlared heartbeat",
                "parameters": [
                    {
                        "description": "Flared heartbeat payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.FlaredHeartbeatPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/flared/ws": {
            "get": {
                "security": [
                    {
                        "TunnelTokenAuth": []
                    }
                ],
                "tags": [
                    "Flared"
                ],
                "summary": "Upgrade OpenFlared connection to websocket",
                "responses": {}
            }
        },
...
        "/api/relay/heartbeat": {
            "post": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Relay"
                ],
                "summary": "Report relay heartbeat",
                "parameters": [
                    {
                        "description": "Relay heartbeat payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.RelayHeartbeatPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/relay/ws": {
            "get": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    }
                ],
                "tags": [
                    "Relay"
                ],
                "summary": "Upgrade relay connection to websocket",
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "controller.DnsAccountInput": {
            "type": "object",
            "properties": {
                "authorization": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "service.AgentNodeHealthEvent": {
            "type": "object",
            "properties": {
                "event_type": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "severity": {
                    "type": "string"
                },
                "triggered_at_unix": {
                    "type": "integer"
                }
            }
        },
        "service.AgentNodeMetricSnapshot": {
            "type": "object",
            "properties": {
                "captured_at_unix": {
                    "type": "integer"
                },
                "cpu_usage_percent": {
                    "type": "number"
                },
                "disk_read_bytes": {
                    "type": "integer"
                },
                "disk_write_bytes": {
                    "type": "integer"
                },
                "memory_total_bytes": {
                    "type": "integer"
                },
                "memory_used_bytes": {
                    "type": "integer"
                },
                "network_rx_bytes": {
                    "type": "integer"
                },
                "network_tx_bytes": {
                    "type": "integer"
                },
                "storage_total_bytes": {
                    "type": "integer"
                },
                "storage_used_bytes": {
                    "type": "integer"
                }
            }
        },
...
        "service.AgentNodeSystemProfile": {
            "type": "object",
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "cpu_cores": {
                    "type": "integer"
                },
                "cpu_model": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "kernel_version": {
                    "type": "string"
                },
                "os_name": {
                    "type": "string"
                },
                "os_version": {
                    "type": "string"
                },
                "reported_at_unix": {
                    "type": "integer"
                },
                "total_disk_bytes": {
                    "type": "integer"
                },
                "total_memory_bytes": {
                    "type": "integer"
                },
                "uptime_seconds": {
                    "type": "integer"
                }
            }
        },
...
        "service.ApplyLogPayload": {
            "type": "object",
            "properties": {
                "checksum": {
                    "type": "string"
                },
                "main_config_checksum": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "node_id": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "route_config_checksum": {
                    "type": "string"
                },
                "support_file_count": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "service.FlaredConnectedRelay": {
            "type": "object",
            "properties": {
                "proxy_count": {
                    "type": "integer"
                },
                "relay_node_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "service.FlaredHeartbeatPayload": {
            "type": "object",
            "properties": {
                "client_version": {
                    "type": "string"
                },
                "connected_relays": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.FlaredConnectedRelay"
                    }
                },
                "current_checksum": {
                    "type": "string"
                },
                "current_version": {
                    "type": "string"
                },
                "frp_version": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "tunnel_status": {
                    "type": "string"
                }
            }
        },
...
        "service.RelayHeartbeatPayload": {
            "type": "object",
            "properties": {
                "frp_version": {
                    "type": "string"
                },
                "frps_client_count": {
                    "type": "integer"
                },
                "frps_connections": {
                    "type": "integer"
                },
                "frps_proxies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.RelayProxyStat"
                    }
                },
                "frps_proxy_count": {
                    "type": "integer"
                },
                "health_events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.AgentNodeHealthEvent"
                    }
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profile": {
                    "$ref": "#/definitions/service.AgentNodeSystemProfile"
                },
                "relay_status": {
                    "type": "string"
                },
                "snapshot": {
                    "$ref": "#/definitions/service.AgentNodeMetricSnapshot"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "service.RelayProxyStat": {
            "type": "object",
            "properties": {
                "client_addr": {
                    "type": "string"
                },
                "client_version": {
                    "type": "string"
                },
                "last_close_time": {
                    "type": "string"
                },
                "last_start_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessTokenAuth": {
            "description": "Agent API 使用节点专属 Agent Token 或全局 Discovery Token",
            "type": "apiKey",
            "name": "X-Agent-Token",
            "in": "header"
        },
        "OpenFlareTokenAuth": {
            "description": "管理端 API 使用登录后返回的用户 Token",
            "type": "apiKey",
            "name": "OPENFLARE_TOKEN",
            "in": "header"
        }
    }
}